requirements.txt
google-genai
python-docx
docxtpl
python-dotenv
pandas
python-jobspy
playwright
pytest-playwright
